QProperty: Force inlining of addObserver
The method is only called in two places: observerProperty (which only
takes care of calling unlink, too, if necessary) and in
registerWithCurrentlyEvaluatingBinding_helper. In the first case, the
method most likely gets inlined anyway. In the latter case, we really
want to avoid the overhead of an additional function call to speed up
registration of the property with the binding. Considering that it is an
internal method, there is no need to worry about code explosion.

Change-Id: I3f0f0e37108f3859321d3b432e37fc8da3c15dc0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 77d528f8d328d43fa367bd4887ad21b5e6b6311e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>