Fix mapping between Han and other CJK scripts
This is a backport of 181bece09afef4e6c3b47e1d246eb7e3ba56d84f
from Qt 6.

There is an overlap between CJK scripts, and fonts that do not
explictly support Han may still have characters from this because
they are used for e.g. Japanese. In addition, these characters
may look slightly different when used in Japanese than in Chinese.

When cc4087b18a488ffb02105d31da6f05acad65b728 was introduced, it
caused a regression on macOS: Since the script of CJK characters
is now retained during font selection and processing, it would
change the order of fallback fonts, causing .AppleJapaneseFont
to end up in the back of the list for Han characters.

The fix is to treat CJK scripts as one when sorting and resolving
fonts.

[ChangeLog][macOS] Fixed an issue where certain shared CJK
characters would be displayed using the Chinese system font
rather than the Japanese one, even on Japanese locale.

Fixes: QTBUG-85598
Change-Id: I050962d739d1292f1c000fd424213b8bd7101e03
Reviewed-by: Andy Shaw <andy.shaw@qt.io>