Fix memory leaks in tst_qlineedit
In the unit tests, two times a validator was used without a parent,
which led to memory leaks. This patch changes this: The
validators are initialized with the widget they are meant for
as parent. This fixes the memory leaks.

Change-Id: I480c0c5104cbe60159fad49df28cbb6240e7ce68
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>
(cherry picked from commit b854bf06dc1902b101c2622f519c84a2a93e2fb1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>